// packages/ui/src/services/api/ApiClient.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BaseHttpRequest } from './core/BaseHttpRequest';
import type { OpenAPIConfig } from './core/OpenAPI';
import { AxiosHttpRequest } from './core/AxiosHttpRequest';

import { ApiKeysService } from './services/ApiKeysService';
import { AuthenticationService } from './services/AuthenticationService';
import { HealthService } from './services/HealthService';
import { ServersService } from './services/ServersService';

type HttpRequestConstructor = new (config: OpenAPIConfig) => BaseHttpRequest;

export class ApiClient {

    public readonly apiKeys: ApiKeysService;
    public readonly authentication: AuthenticationService;
    public readonly health: HealthService;
    public readonly servers: ServersService;

    public readonly request: BaseHttpRequest;

    constructor(config?: Partial<OpenAPIConfig>, HttpRequest: HttpRequestConstructor = AxiosHttpRequest) {
        this.request = new HttpRequest({
            BASE: config?.BASE ?? 'http://localhost:3010',
            VERSION: config?.VERSION ?? '1.0.0',
            WITH_CREDENTIALS: config?.WITH_CREDENTIALS ?? false,
            CREDENTIALS: config?.CREDENTIALS ?? 'include',
            TOKEN: config?.TOKEN,
            USERNAME: config?.USERNAME,
            PASSWORD: config?.PASSWORD,
            HEADERS: config?.HEADERS,
            ENCODE_PATH: config?.ENCODE_PATH,
        });

        this.apiKeys = new ApiKeysService(this.request);
        this.authentication = new AuthenticationService(this.request);
        this.health = new HealthService(this.request);
        this.servers = new ServersService(this.request);
    }
}

